samaisite/requirements.txt
Django==4.0.3
# PyCryptodome==3.12.0
# pyjwt==1.7.1
# python-dateutil==2.8.1; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.5"
# python-dotenv==0.14.0
# python-json-logger==2.0.1; python_version >= "3.4"
# pytimeparse==1.1.8
# urllib3==1.5

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0

# # Logging -------------------------------------
# tensorboard>=2.4.1
# # wandb

# # Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0
thop